#176 index into glif example script, allowing common variables
