--- doc_template/examples_root/examples/glif_ex5.py
@@ -1,52 +1,0 @@
-import numpy as np
-import matplotlib.pyplot as plt
-
-voltage = output['voltage']
-threshold = output['threshold']
-interpolated_spike_times = output['interpolated_spike_times']
-spike_times = output['interpolated_spike_times']
-interpolated_spike_voltages = output['interpolated_spike_voltage']
-interpolated_spike_thresholds = output['interpolated_spike_threshold']
-grid_spike_indices = output['spike_time_steps']
-grid_spike_times = output['grid_spike_times']
-after_spike_currents = output['AScurrents']
-
-# create a time array for plotting
-time = np.arange(len(stimulus))*neuron.dt
-
-plt.figure(figsize=(10, 10))
-
-# plot stimulus
-plt.subplot(3,1,1)
-plt.plot(time, stimulus)
-plt.xlabel('time (s)')
-plt.ylabel('current (A)')
-plt.title('Stimulus')
-
-# plot model output
-plt.subplot(3,1,2)
-plt.plot(time,  voltage, label='voltage')
-plt.plot(time,  threshold, label='threshold')
-
-if grid_spike_indices is not None:
-    plt.plot(interpolated_spike_times, interpolated_spike_voltages, 'x', 
-             label='interpolated spike')
-
-    plt.plot((grid_spike_indices-1)*neuron.dt, voltage[grid_spike_indices-1], '.', 
-             label='last step before spike')
-
-plt.xlabel('time (s)')
-plt.ylabel('voltage (V)')
-plt.legend(loc=3)
-plt.title('Model Response')
-
-# plot after spike currents
-plt.subplot(3,1,3)
-for ii in range(np.shape(after_spike_currents)[1]):
-    plt.plot(time, after_spike_currents[:,ii])
-plt.xlabel('time (s)')
-plt.ylabel('current (A)')
-plt.title('After Spike Currents')
-
-plt.tight_layout()
-plt.show()
