--- doc_template/examples_root/examples/glif_ex6.py
@@ -1,15 +1,0 @@
-# define your own custom voltage reset rule 
-# this one linearly scales the input voltage
-def custom_voltage_reset_rule(neuron, voltage_t0, custom_param_a, custom_param_b):
-    return custom_param_a * voltage_t0 + custom_param_b
-
-# initialize a neuron from a neuron config file
-neuron_config = json_utilities.read('neuron_config.json')
-neuron = GlifNeuron.from_dict(neuron_config)
-
-# configure a new method and overwrite the neuron's old method
-method = neuron.configure_method('custom', custom_voltage_reset_rule, 
-                                 { 'custom_param_a': 0.1, 'custom_param_b': 0.0 })
-neuron.voltage_reset_method = method
-
-output = neuron.run(stimulus)
